# BE/swagger.yaml
openapi: 3.0.0
info:
  title: Noctua Owls API
  description: API para la gestión del equipo de E-Sports Noctua Owls (Universidad Nacional de Colombia).
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Servidor de Desarrollo
  - url: https://noctuaowls.onrender.com/api
    description: Servidor de Producción
...
paths:
...
  /bd/usuarios:
    get:
      summary: Obtener listado de usuarios y staff
      tags:
        - BD
      parameters:
        - in: query
          name: juegos
          schema:
            type: array
            items:
              type: string
          description: Lista de IDs de juegos para filtrar
          style: form
          explode: false
        - in: query
          name: cargos
          schema:
            type: array
            items:
              type: string
          description: Lista de cargos para filtrar
          style: form
          explode: false
      responses:
        "200":
          description: Lista de usuarios filtrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
...